# services/exam-service/docs/swagger/swagger.yaml
basePath: /
definitions:
  DifficultyLevel:
    enum:
    - Easy
    - Medium
    - Hard
    type: string
    x-enum-varnames:
    - DifficultyEasy
    - DifficultyMedium
    - DifficultyHard
...
  Question:
    properties:
      content:
        type: string
      createdAt:
        type: string
      createdBy:
        type: string
      difficulty:
        $ref: '#/definitions/DifficultyLevel'
      enterpriseId:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      mediaUrl:
        type: string
      metadata:
        additionalProperties: true
        type: object
      negativePoints:
        type: number
      options:
        items:
          $ref: '#/definitions/QuestionOption'
        type: array
      points:
        type: integer
      title:
        type: string
      topic:
        type: string
      type:
        $ref: '#/definitions/QuestionType'
      updatedAt:
        type: string
    type: object
  QuestionOption:
    properties:
      content:
        type: string
      id:
        type: string
      isCorrect:
        type: boolean
      questionId:
        type: string
    type: object
  QuestionType:
    enum:
    - MCQ
    - TrueFalse
    - ShortAnswer
    - Essay
    type: string
    x-enum-varnames:
    - QuestionTypeMCQ
    - QuestionTypeTrueFalse
    - QuestionTypeShortAnswer
    - QuestionTypeEssay
...
  ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact:
    name: Veritas Platform Team
  description: Exam authoring, scheduling, publishing, and question bank management
    service.
  title: Veritas Exam Service API
  version: "1.0"
paths:
...
  /questions:
    get:
      description: List all questions for the caller enterprise.
      parameters:
      - description: Enterprise ID (UUID)
        in: header
        name: X-Enterprise-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/Question'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: List questions
      tags:
      - question
    post:
      consumes:
      - application/json
      description: Create one question under the caller enterprise.
      parameters:
      - description: Enterprise ID (UUID)
        in: header
        name: X-Enterprise-ID
        required: true
        type: string
      - description: Actor user ID (UUID)
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Question payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/Question'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Question'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Create question
      tags:
      - question
  /questions/{questionId}:
    delete:
      description: Delete one question from the caller enterprise.
      parameters:
      - description: Enterprise ID (UUID)
        in: header
        name: X-Enterprise-ID
        required: true
        type: string
      - description: Question ID (UUID)
        in: path
        name: questionId
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Delete question
      tags:
      - question
    get:
      description: Get a question by ID for the caller enterprise.
      parameters:
      - description: Enterprise ID (UUID)
        in: header
        name: X-Enterprise-ID
        required: true
        type: string
      - description: Question ID (UUID)
        in: path
        name: questionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Question'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get question
      tags:
      - question
    patch:
      consumes:
      - application/json
      description: Update question fields by ID.
      parameters:
      - description: Enterprise ID (UUID)
        in: header
        name: X-Enterprise-ID
        required: true
        type: string
      - description: Actor user ID (UUID)
        in: header
        name: X-User-ID
        required: true
        type: string
      - description: Question ID (UUID)
        in: path
        name: questionId
        required: true
        type: string
      - description: Question payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/Question'
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Update question
      tags:
      - question
schemes:
- http
- https
swagger: "2.0"
...